- CS_Build_v0.1 (master) -> [ Managed to pass structure arg into UDP thread handler, able to reply back to client on 2 separate threads (same local IP, different ports - minimal working server reply using separate threads - added static uint array and enum for Server Ports (as in IP's) to refactor struct arg instantiation objects in a for-loop - commit #5 - 112021 ] | SERVER PROD BUILDS = [ Refactored build for DEBUG MODE (DBG MACRO) ] RPi4 Release Backup is included - Server_vETHx instances are RPi4 RELEASE-READY - NEEDS TESTING - NON-LOCAL - Commit #8 - 111921

+++ Inter-Process-Communication_Testing/CS_Build_v0.1/Server_v0.1/inc/CS_Common.h
@@ -96,7 +96,7 @@ enum
   eREM_SRV_IP_5
 };
 
-static uint8_t uRem_Srv_Port[] =
+static uint16_t uRem_Srv_Port[] =
 {
   REM_SRV_PORT_0,
   REM_SRV_PORT_1,
